fix(mobile/ios): system font loading, CMTime encoding, ATT status type, default stderr logger
- text_system: font-kit's Font::properties() unwraps numeric CoreText traits that iOS
  omits for many system families, aborting the process the first time a non-default
  font family (e.g. Menlo) is requested. Read symbolic traits directly instead and
  only require the symbolic trait when loading a family.
- video_player/audio: CMTime is an anonymous-struct typedef; its ObjC encoding is
  {?=qiIq}, so naming it "CMTime" made objc2's encoding verification abort on
  every currentTime/duration call.
- permission_handler: ATTrackingManager.trackingAuthorizationStatus is NSUInteger.
- ffi: install a stderr logger in gpui_ios_initialize when the host has none, so
  log output reaches the Xcode console instead of being dropped.

+++ crates/gpui_mobile/src/packages/permission_handler/ios.rs
@@ -188,7 +188,7 @@ unsafe fn check_speech_authorization() -> Result<PermissionStatus, String> {
 }
 
 unsafe fn check_tracking_authorization() -> Result<PermissionStatus, String> {
-    let status: u32 = msg_send![class!(ATTrackingManager), trackingAuthorizationStatus];
+    let status: usize = msg_send![class!(ATTrackingManager), trackingAuthorizationStatus];
     Ok(match status {
         0 => PermissionStatus::Denied, // ATTrackingManagerAuthorizationStatusNotDetermined
         1 => PermissionStatus::Restricted, // ATTrackingManagerAuthorizationStatusRestricted
